Moved Chat to a module folder, moved ChatData class to the globals
