+++ lib/src/bot.dart
@@ -1,5 +1,4 @@
 import 'package:weather/src/core/config.dart';
-import 'package:weather/src/core/chat.dart';
 import 'package:weather/src/core/user.dart';
 
 import 'package:weather/src/globals/bot_command.dart';
@@ -7,7 +6,8 @@ import 'package:weather/src/globals/access_level.dart';
 import 'package:weather/src/injector/injection.dart';
 import 'package:weather/src/platform/platform.dart';
 
-import 'package:weather/src/modules/chat_manager.dart';
+import 'package:weather/src/modules/chat/chat.dart';
+import 'package:weather/src/modules/chat/chat_manager.dart';
 import 'package:weather/src/modules/user_manager.dart';
 import 'package:weather/src/modules/weather_manager.dart';
 import 'package:weather/src/modules/panorama_manager.dart';
