+++ lib/src/modules/accordion_poll/accordion_poll_manager.dart
@@ -1,5 +1,5 @@
 import 'package:weather/src/core/user.dart';
-import 'package:weather/src/core/chat.dart';
+import 'package:weather/src/modules/chat/chat.dart';
 import 'package:weather/src/globals/chat_platform.dart';
 import 'package:weather/src/globals/message_event.dart';
 import 'package:weather/src/platform/platform.dart';
