+++ lib/src/modules/general_manager.dart
@@ -1,7 +1,7 @@
-import 'package:weather/src/core/chat.dart';
 import 'package:weather/src/core/general.dart';
 import 'package:weather/src/globals/message_event.dart';
 import 'package:weather/src/platform/platform.dart';
+import 'package:weather/src/modules/chat/chat.dart';
 import 'package:weather/src/modules/utils.dart';
 
 class GeneralManager {
