+++ test/chat_test.dart
@@ -1,7 +1,7 @@
 import 'package:postgres/postgres.dart';
 import 'package:test/test.dart';
 import 'package:weather/src/globals/chat_platform.dart';
-import 'package:weather/src/core/chat.dart';
+import 'package:weather/src/modules/chat/chat.dart';
 import 'utils/setup.dart';
 import 'utils/db_connection.dart';
 import 'utils/helpers.dart';
